services/hl7-processor/requirements.txt
# HL7 Processor Service Requirements
# Core HL7 processing libraries
hl7apy==1.3.4
hl7==0.4.5

# Web framework
fastapi==0.95.2
uvicorn[standard]==0.22.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Message queue
celery==5.3.4
redis==5.0.1

# HTTP client
httpx==0.25.2
requests==2.31.0

# Validation and serialization
pydantic==1.10.15
marshmallow==3.20.1

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Security
cryptography==46.0.1
pyjwt==2.8.0

# Configuration
python-decouple==3.8
pyyaml==6.0.1

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Healthcare standards
fhir.resources==7.0.2
pydicom==2.4.3
